requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
blis==0.4.1
bs4==0.0.1
catalogue==1.0.0
certifi==2020.11.8
cffi @ file:///D:/bld/cffi_1606601395031/work
chardet==3.0.4
click==7.1.2
confusable-homoglyphs==3.2.0
cryptography==3.2.1
cymem==2.0.4
dataclasses==0.6
defusedxml==0.6.0
Django @ file:///tmp/build/80754af9/django_1601563330930/work
django-allauth==0.44.0
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-extra-views==0.13.0
django-humanize==0.1.2
django-registration==3.1.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
Faker==4.18.0
filelock==3.0.12
future==0.18.2
futures==3.1.1
goslate==1.5.1
humanize==3.1.0
idna==2.10
importlib-metadata==2.0.0
jdcal==1.4.1
joblib==0.17.0
misaka @ file:///D:/bld/misaka_1602586490060/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
murmurhash==1.0.4
mysqlclient==2.0.1
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
oauthlib==3.1.0
olefile==0.46
openpyxl @ file:///tmp/build/80754af9/openpyxl_1598113097404/work
packaging==20.4
Pillow @ file:///C:/ci/pillow_1603823068645/work
plac==1.1.3
preshed==3.0.4
protobuf==3.14.0
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
PyDictionary==2.0.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-resize-image==1.1.19
python3-openid==3.2.0
pytz==2020.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
sacremoses==0.0.43
sentencepiece==0.1.91
six @ file:///C:/ci/six_1605187374963/work
sorl-thumbnail==12.7.0
soupsieve==2.0.1
spacy==2.3.4
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
srsly==1.0.4
text-unidecode==1.3
thinc==7.4.1
tokenizers==0.9.3
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tqdm==4.54.0
transformers==3.5.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.26.2
wasabi==0.8.0
wincertstore==0.2
zipp==3.4.0
